Refactored list/ to storage/ in prep for storing prices
